# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "Brief": {
        "description": "El encargo de la obra.\n\n`start` es el instante en que arranca el mundo: todo lo que el brief\nestablece es cierto desde ahi, y nada anterior existe.\n\nLo que va despues de `word_tolerance` lo trae la entrevista (RF-200). Es\nopcional para que un brief escrito a mano, sin destinatario, siga valiendo.",
        "properties": {
          "dedication": {
            "default": "",
            "title": "Dedication",
            "type": "string"
          },
          "entities": {
            "items": {
              "$ref": "#/components/schemas/BriefEntity"
            },
            "title": "Entities",
            "type": "array"
          },
          "forbidden_themes": {
            "items": {
              "type": "string"
            },
            "title": "Forbidden Themes",
            "type": "array"
          },
          "forbidden_words": {
            "items": {
              "type": "string"
            },
            "title": "Forbidden Words",
            "type": "array"
          },
          "genre": {
            "default": "",
            "title": "Genre",
            "type": "string"
          },
          "recipient": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Recipient"
              },
              {
                "type": "null"
              }
            ]
          },
          "relations": {
            "items": {
              "$ref": "#/components/schemas/BriefRelation"
            },
            "title": "Relations",
            "type": "array"
          },
          "rulebook": {
            "default": "",
            "description": "DEP-02, vacio si no es deportiva",
            "title": "Rulebook",
            "type": "string"
          },
          "start": {
            "$ref": "#/components/schemas/WorldTime"
          },
          "style_guide": {
            "description": "POE-06",
            "minLength": 1,
            "title": "Style Guide",
            "type": "string"
          },
          "target_words": {
            "exclusiveMinimum": 0.0,
            "title": "Target Words",
            "type": "integer"
          },
          "title": {
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "tone": {
            "default": "",
            "title": "Tone",
            "type": "string"
          },
          "word_tolerance": {
            "default": 0.1,
            "exclusiveMaximum": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Word Tolerance",
            "type": "number"
          }
        },
        "required": [
          "title",
          "start",
          "entities",
          "style_guide",
          "target_words"
        ],
        "title": "Brief",
        "type": "object"
      },
...
      "ChangeRequestIn": {
        "description": "RI-47. La peticion y su ancla: un fragmento de la lectura o un hecho de la ficha.",
        "properties": {
          "anchor": {
            "discriminator": {
              "mapping": {
                "fact": "#/components/schemas/FactAnchor",
                "fragment": "#/components/schemas/FragmentAnchor"
              },
              "propertyName": "kind"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/FragmentAnchor"
              },
              {
                "$ref": "#/components/schemas/FactAnchor"
              }
            ],
            "title": "Anchor"
          },
          "text": {
            "description": "Texto no confiable",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text",
          "anchor"
        ],
        "title": "ChangeRequestIn",
        "type": "object"
      },
      "ChangeRequestList": {
        "properties": {
          "requests": {
            "items": {
              "$ref": "#/components/schemas/ChangeRequest"
            },
            "title": "Requests",
            "type": "array"
          }
        },
        "required": [
          "requests"
        ],
        "title": "ChangeRequestList",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NovelCreated": {
        "properties": {
          "events_loaded": {
            "description": "Hechos que el brief aporto al canon",
            "title": "Events Loaded",
            "type": "integer"
          },
          "novel_id": {
            "title": "Novel Id",
            "type": "string"
          }
        },
        "required": [
          "novel_id",
          "events_loaded"
        ],
        "title": "NovelCreated",
        "type": "object"
      },
...
      "NovelList": {
        "properties": {
          "novels": {
            "items": {
              "$ref": "#/components/schemas/NovelEntry"
            },
            "title": "Novels",
            "type": "array"
          }
        },
        "required": [
          "novels"
        ],
        "title": "NovelList",
        "type": "object"
      },
...
      "RunState": {
        "description": "RI-03. El estado de la tirada, solo lo que un lector puede ver.",
        "properties": {
          "chapter_in_progress": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chapter In Progress"
          },
          "closed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "description": "`None` mientras la tirada no ha terminado",
            "title": "Closed"
          },
          "error": {
            "default": "",
            "title": "Error",
            "type": "string"
          },
          "frozen_chapters": {
            "minimum": 0.0,
            "title": "Frozen Chapters",
            "type": "integer"
          },
          "last_closed_scene": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Closed Scene"
          },
          "novel_id": {
            "title": "Novel Id",
            "type": "string"
          },
          "quarantines": {
            "description": "Veces que un capitulo se rehizo entero",
            "minimum": 0.0,
            "title": "Quarantines",
            "type": "integer"
          },
          "reason": {
            "default": "",
            "title": "Reason",
            "type": "string"
          },
          "running": {
            "title": "Running",
            "type": "boolean"
          }
        },
        "required": [
          "novel_id",
          "running",
          "frozen_chapters",
          "quarantines"
        ],
        "title": "RunState",
        "type": "object"
      },
...
      "canon__routes__ErrorDetail": {
        "description": "Cuerpo de todo error de esta API.\n\nSe declara porque un contrato que dice \"puede devolver 404\" sin decir con\nque forma deja al cliente generado adivinando, que es la mitad del problema\nque el contrato existe para quitar.",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "orchestration__routes__ErrorDetail": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Sistema autonomo de generacion de novelas largas. La API observa y arranca; no aprueba, no corrige y no desbloquea.",
    "title": "Story Maker",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/novels": {
      "get": {
        "description": "RI-37. Las novelas del directorio de tiradas, con su estado y sus versiones.\n\nVive aqui y no en `canon/` porque el estado de la tirada es del Orquestador.",
        "operationId": "list_novels_novels_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NovelList"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Novels",
        "tags": [
          "orchestration"
        ]
      },
      "post": {
        "description": "RI-01. Carga el brief. Es la unica escritura que la API provoca.",
        "operationId": "create_novels_post",
        "parameters": [
          {
            "description": "Identificador de la novela",
            "in": "query",
            "name": "novel_id",
            "required": true,
            "schema": {
              "description": "Identificador de la novela",
              "pattern": "^[a-z0-9][a-z0-9-]{2,63}$",
              "title": "Novel Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Brief"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NovelCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/canon__routes__ErrorDetail"
                }
              }
            },
            "description": "Bad Request"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/canon__routes__ErrorDetail"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create",
        "tags": [
          "canon"
        ]
      }
    },
    "/novels/{novel_id}": {
      "get": {
        "description": "RI-03. Capitulo y escena en curso, congelados, cuarentenas, cierre.",
        "operationId": "get_state_novels__novel_id__get",
        "parameters": [
          {
            "description": "Identificador de la novela",
            "in": "path",
            "name": "novel_id",
            "required": true,
            "schema": {
              "description": "Identificador de la novela",
              "pattern": "^[a-z0-9][a-z0-9-]{2,63}$",
              "title": "Novel Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunState"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/orchestration__routes__ErrorDetail"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/orchestration__routes__ErrorDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get State",
        "tags": [
          "orchestration"
        ]
      }
    },
    "/novels/{novel_id}/change-requests": {
      "get": {
        "description": "RI-48. Todas las solicitudes con su estado.",
        "operationId": "list_change_requests_novels__novel_id__change_requests_get",
        "parameters": [
          {
            "description": "Identificador de la novela",
            "in": "path",
            "name": "novel_id",
            "required": true,
            "schema": {
              "description": "Identificador de la novela",
              "pattern": "^[a-z0-9][a-z0-9-]{2,63}$",
              "title": "Novel Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeRequestList"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/orchestration__routes__ErrorDetail"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/orchestration__routes__ErrorDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Change Requests",
        "tags": [
          "orchestration"
        ]
      },
      "post": {
        "description": "RI-47. Interpreta en la misma llamada (D-82) y deja la solicitud en cola o rechazada.\n\nCon la tirada en marcha, el bucle la aplica tras el siguiente capitulo\ncongelado. Sin tirada, el Orquestador la aplica enseguida en su hilo (D-77).",
        "operationId": "create_change_request_novels__novel_id__change_requests_post",
        "parameters": [
          {
            "description": "Identificador de la novela",
            "in": "path",
            "name": "novel_id",
            "required": true,
            "schema": {
              "description": "Identificador de la novela",
              "pattern": "^[a-z0-9][a-z0-9-]{2,63}$",
              "title": "Novel Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeRequestIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeRequest"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/orchestration__routes__ErrorDetail"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/orchestration__routes__ErrorDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Change Request",
        "tags": [
          "orchestration"
        ]
      }
    },
    "/novels/{novel_id}/change-requests/{request_id}": {
      "get": {
        "description": "RI-49. Estado, interpretacion y, si se aplico, la version que produjo.",
        "operationId": "get_change_request_novels__novel_id__change_requests__request_id__get",
        "parameters": [
          {
            "description": "Identificador de la novela",
            "in": "path",
            "name": "novel_id",
            "required": true,
            "schema": {
              "description": "Identificador de la novela",
              "pattern": "^[a-z0-9][a-z0-9-]{2,63}$",
              "title": "Novel Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Request Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeRequest"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/orchestration__routes__ErrorDetail"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/orchestration__routes__ErrorDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Change Request",
        "tags": [
          "orchestration"
        ]
      }
    },
...
    "/novels/{novel_id}/run": {
      "post": {
        "description": "RI-02. Arranca la tirada. Idempotente: si ya corre, devuelve el estado.",
        "operationId": "start_run_novels__novel_id__run_post",
        "parameters": [
          {
            "description": "Identificador de la novela",
            "in": "path",
            "name": "novel_id",
            "required": true,
            "schema": {
              "description": "Identificador de la novela",
              "pattern": "^[a-z0-9][a-z0-9-]{2,63}$",
              "title": "Novel Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunState"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/orchestration__routes__ErrorDetail"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/orchestration__routes__ErrorDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Run",
        "tags": [
          "orchestration"
        ]
      }
    },
...
    "/novels/{novel_id}/trace": {
      "get": {
        "description": "RI-27. Los registros de la traza, en orden de escritura.",
        "operationId": "get_trace_novels__novel_id__trace_get",
        "parameters": [
          {
            "description": "Identificador de la novela",
            "in": "path",
            "name": "novel_id",
            "required": true,
            "schema": {
              "description": "Identificador de la novela",
              "pattern": "^[a-z0-9][a-z0-9-]{2,63}$",
              "title": "Novel Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 500,
              "maximum": 5000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 40,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TracePage"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/orchestration__routes__ErrorDetail"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/orchestration__routes__ErrorDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Trace",
        "tags": [
          "orchestration"
        ]
      }
    },
...
  }
}